webapp/requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
openpiv>=0.25.0
Pillow>=10.0.0
screeninfo>=0.8.1
